# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My OpenApi Documentation",
        "description": "My OpenApi description",
        "contact": {
            "email": "markusmak@gmail.com"
        },
        "license": {
            "name": "Proprietary",
            "url": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "My API Server"
        }
    ],
    "paths": {
...
        "/api/signature": {
            "get": {
                "tags": [
                    "Signatures"
                ],
                "summary": "Exibe uma lista dos recursos de assinatura.",
                "operationId": "8f40e92c51a4e10e6bc4f906b0cb1393",
                "responses": {
                    "200": {
                        "description": "Retorna a lista de assinaturas com sucesso."
                    },
                    "400": {
                        "description": "Retorna um erro se ocorrer um problema durante a busca das assinaturas."
                    }
                }
            },
            "post": {
                "tags": [
                    "Signatures"
                ],
                "summary": "Armazena um novo recurso de assinatura.",
                "operationId": "65f6c4465b51f4296abdd8a4d7e8efd6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignatureRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna os dados da nova assinatura criada com sucesso."
                    },
                    "400": {
                        "description": "Retorna um erro se ocorrer um problema durante a criação da assinatura."
                    }
                }
            }
        },
        "/api/signature/{id}": {
            "get": {
                "tags": [
                    "Signatures"
                ],
                "summary": "Exibe o recurso de assinatura especificado.",
                "operationId": "6a17388c026d20166c9eb232cc4fc384",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da assinatura a ser recuperada",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os detalhes da assinatura encontrada com sucesso."
                    },
                    "400": {
                        "description": "Retorna um erro se a assinatura não for encontrada."
                    }
                }
            },
            "put": {
                "tags": [
                    "Signatures"
                ],
                "summary": "Atualiza o recurso de assinatura especificado.",
                "operationId": "bed9dc54866c2e0a9568feefe93bcfe8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da assinatura a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignatureRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna os detalhes da assinatura atualizada com sucesso."
                    },
                    "400": {
                        "description": "Retorna um erro se a assinatura não for encontrada."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Signatures"
                ],
                "summary": "Exclui o recurso de assinatura especificado.",
                "operationId": "0663d3e44b4a00842afa8b8ab15eba65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da assinatura a ser excluída",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna uma mensagem de sucesso após a exclusão."
                    },
                    "400": {
                        "description": "Retorna um erro se a assinatura não for encontrada."
                    }
                }
            }
        },
        "/api/verification_of_signatures": {
            "get": {
                "tags": [
                    "Signatures"
                ],
                "summary": "Função para gerar as faturas.",
                "operationId": "fae467bbadcbe95d2a9b7dc68157e15f",
                "responses": {
                    "200": {
                        "description": "Função para gerar as faturas."
                    },
                    "400": {
                        "description": "Retorna um erro se ocorrer um problema durante a busca das assinaturas."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SignatureRequest": {
                "title": "SignatureRequest",
                "description": "Request schema for Signature",
                "required": [
                    "cadastre_id",
                    "describe",
                    "value"
                ],
                "properties": {
                    "cadastre_id": {
                        "description": "ID of the related cadastre",
                        "type": "integer",
                        "example": 1
                    },
                    "describe": {
                        "description": "Description of the signature",
                        "type": "string",
                        "example": "Signature description"
                    },
                    "value": {
                        "description": "Value of the signature",
                        "type": "string",
                        "example": "Signature value"
                    }
                },
                "type": "object"
            }
        }
    }
}